feat(#251): Refactor ripgrep-search extension — split 806-line monolith into module directory
