+++ benchmarks/benchmark-tools.sh
@@ -18,7 +18,7 @@ CONFIGS=(
   "1-no-tools|--no-tools --no-extensions"
   "2-builtin-mapper|--no-extensions -e $PROJECT_DIR/.pi/extensions/ranked-map.ts"
   "3-builtin-mapper-structural|--no-extensions -e $PROJECT_DIR/.pi/extensions/ranked-map.ts -e $PROJECT_DIR/.pi/extensions/structural-analyzer.ts"
-  "4-builtin-mapper-structural-rg|--no-extensions -e $PROJECT_DIR/.pi/extensions/ranked-map.ts -e $PROJECT_DIR/.pi/extensions/structural-analyzer.ts -e $PROJECT_DIR/.pi/extensions/ripgrep-search.ts"
+  "4-builtin-mapper-structural-rg|--no-extensions -e $PROJECT_DIR/.pi/extensions/ranked-map.ts -e $PROJECT_DIR/.pi/extensions/structural-analyzer.ts -e $PROJECT_DIR/.pi/extensions/ripgrep-search/index.ts"
 )
 
 GREEN='\033[0;32m'
